feat: 가까운 엣지 계산 최적화
